requirements.txt
# 核心框架
llama-factory>=0.8.0
transformers>=4.41.0
torch>=2.1.0
accelerate>=0.30.0

# 图像处理
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# 配置和工具
PyYAML>=6.0
tqdm>=4.66.0
numpy>=1.24.0
